+++ providers/dotsql/sql.go
@@ -1,22 +1,78 @@
-package xtemplate
+package sql
 
 import (
 	"context"
 	"database/sql"
+	"errors"
 	"fmt"
 	"iter"
 	"log/slog"
 	"text/template"
 	"time"
+
+	"github.com/infogulch/xtemplate"
 )
 
-// DotDB is used to create a dot field value that can query a SQL database. When
+func init() {
+	xtemplate.Register("sql", func() xtemplate.DotConfig { return &DotSqlConfig{} })
+}
+
+func WithSql(name string, db *sql.DB, opt *sql.TxOptions) xtemplate.Option {
+	return func(c *xtemplate.Config) error {
+		if db == nil {
+			return fmt.Errorf("cannot create database provider with nil sql.DB. name: %s", name)
+		}
+		c.Providers = append(c.Providers, &DotSqlConfig{Name: name, DB: db, TxOptions: opt})
+		return nil
+	}
+}
+
+type DotSqlConfig struct {
+	*sql.DB        `json:"-"`
+	*sql.TxOptions `json:"-"`
+	Name           string `json:"name"`
+	Driver         string `json:"driver"`
+	Connstr        string `json:"connstr"`
+	MaxOpenConns   int    `json:"max_open_conns"`
+}
+
+var _ xtemplate.CleanupDotProvider = &DotSqlConfig{}
+
+func (d *DotSqlConfig) FieldName() string { return d.Name }
+func (d *DotSqlConfig) Init(ctx context.Context) error {
+	if d.DB != nil {
+		return nil
+	}
+	db, err := sql.Open(d.Driver, d.Connstr)
+	if err != nil {
+		return fmt.Errorf("failed to open database with driver name '%s': %w", d.Driver, err)
+	}
+	db.SetMaxOpenConns(d.MaxOpenConns)
+	if err := db.Ping(); err != nil {
+		return fmt.Errorf("failed to ping database on open: %w", err)
+	}
+	d.DB = db
+	return nil
+}
+func (d *DotSqlConfig) Value(r xtemplate.Request) (any, error) {
+	return &DotSql{d.DB, xtemplate.GetLogger(r.R.Context()), r.R.Context(), d.TxOptions, nil}, nil
+}
+func (dp *DotSqlConfig) Cleanup(v any, err error) error {
+	d := v.(*DotSql)
+	if err != nil {
+		return errors.Join(err, d.rollback())
+	} else {
+		return errors.Join(err, d.commit())
+	}
+}
+
+// DotSql is used to create a dot field value that can query a SQL database. When
 // any of its statement executing methods are called, it creates a new
 // transaction. When template execution finishes, if there were no errors it
 // automatically commits any uncommitted transactions remaining after template
 // execution completes, but if there were errors then it calls rollback on the
 // transaction.
-type DotDB struct {
+type DotSql struct {
 	db  *sql.DB
 	log *slog.Logger
 	ctx context.Context
@@ -24,7 +80,7 @@ type DotDB struct {
 	tx  *sql.Tx
 }
 
-func (d *DotDB) makeTx() (err error) {
+func (d *DotSql) makeTx() (err error) {
 	if d.tx == nil {
 		d.tx, err = d.db.BeginTx(d.ctx, d.opt)
 	}
@@ -34,7 +90,7 @@ func (d *DotDB) makeTx() (err error) {
 // Exec executes a statement with parameters and returns the raw [sql.Result].
 // Note: this can be a bit difficult to use inside a template, consider using
 // other methods that provide easier to use return values.
-func (c *DotDB) Exec(query string, params ...any) (result sql.Result, err error) {
+func (c *DotSql) Exec(query string, params ...any) (result sql.Result, err error) {
 	if err = c.makeTx(); err != nil {
 		return
 	}
@@ -57,7 +113,7 @@ func (c *DotDB) Exec(query string, params ...any) (result sql.Result, err error)
 // via panic(template.ExecError{...}); the template engine recovers these and
 // returns them from template execution, aborting it just like a normal error
 // return would.
-func (c *DotDB) QueryRows(query string, params ...any) (iter.Seq[map[string]any], error) {
+func (c *DotSql) QueryRows(query string, params ...any) (iter.Seq[map[string]any], error) {
 	if err := c.makeTx(); err != nil {
 		return nil, err
 	}
@@ -113,7 +169,7 @@ func (c *DotDB) QueryRows(query string, params ...any) (iter.Seq[map[string]any]
 
 // QueryRow executes a query, which must return one row, and returns it as a
 // map[string]any.
-func (c *DotDB) QueryRow(query string, params ...any) (map[string]any, error) {
+func (c *DotSql) QueryRow(query string, params ...any) (map[string]any, error) {
 	rows, err := c.QueryRows(query, params...)
 	if err != nil {
 		return nil, err
@@ -135,7 +191,7 @@ func (c *DotDB) QueryRow(query string, params ...any) (map[string]any, error) {
 
 // QueryVal executes a query, which must return one row with one column, and
 // returns the value of the column.
-func (c *DotDB) QueryVal(query string, params ...any) (any, error) {
+func (c *DotSql) QueryVal(query string, params ...any) (any, error) {
 	row, err := c.QueryRow(query, params...)
 	if err != nil {
 		return nil, err
@@ -149,14 +205,14 @@ func (c *DotDB) QueryVal(query string, params ...any) (any, error) {
 	panic("impossible condition")
 }
 
-// Commit manually commits any implicit transactions opened by this DotDB. This
+// Commit manually commits any implicit transactions opened by this DotSql. This
 // is called automatically if there were no errors at the end of template
 // execution.
-func (c *DotDB) Commit() (string, error) {
+func (c *DotSql) Commit() (string, error) {
 	return "", c.commit()
 }
 
-func (c *DotDB) commit() error {
+func (c *DotSql) commit() error {
 	if c.tx != nil {
 		err := c.tx.Commit()
 		c.log.Debug("commit", slog.Any("error", err))
@@ -166,14 +222,14 @@ func (c *DotDB) commit() error {
 	return nil
 }
 
-// Rollback manually rolls back any implicit transactions opened by this DotDB.
+// Rollback manually rolls back any implicit transactions opened by this DotSql.
 // This is called automatically if there were any errors that occurred during
-// template exeuction.
-func (c *DotDB) Rollback() (string, error) {
+// template execution.
+func (c *DotSql) Rollback() (string, error) {
 	return "", c.rollback()
 }
 
-func (c *DotDB) rollback() error {
+func (c *DotSql) rollback() error {
 	if c.tx != nil {
 		err := c.tx.Rollback()
 		c.log.Debug("rollback", slog.Any("error", err))